requirements.txt
python-dateutil==2.8.2
requests==2.28.2
